+++ CHANGELOG.md
@@ -1,10 +1,23 @@
+# 1.3.0
+## Added
+- Quantity options to set the quantity for each item in the merchant
+- Amount options to set the amount of returned items in the merchant
+- Premade Shops Compendium
+    - Runes Merchants
+- Setup for additional compendiums
+- Excluded more non-obtainable items
+- Changed the folder structure to try to seem more profesional
+---
+---
 # 1.2.0
 ## Added
 - Setup to Exclude non-obtainable items (Ancestry/Class items)
 ---
+---
 # 1.1.0
 ## Added
 - PF2E Toolbelt - Better Merchant setup
+---
 ---
 # 1.0.0
 ## Added